requirements.txt
blinker==1.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
Flask-BasicAuth==0.2.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sentry-sdk==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.6
Werkzeug==2.0.1
